packages/gcp-executor/requirements.txt
flask==2.3.3
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
edge-tts>=6.1.9
Pillow>=10.0.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
moviepy>=1.0.3